Adding and Fetching Player from server.
The Signup page now creates users onto the backend, and the ladder
page displays the users from the server. The ladder is updated upon
clicking and entering the ladder page as well as peroidically asking
the backend to resend the user list.

Api calls use promises to maintain validity, currently the reject
path does not give adequate error information to display to the
end user.